added download qr image in users view and updated delete user,logs, and records

+++ controllers/adminController.js
@@ -14,29 +14,22 @@ const Record = require('../models/records');
 module.exports.admin_get = (req, res) => {
     User.countDocuments({}, function(err, count) {
         let employees = count;
-        console.log("Employees: " + employees);
-
         User.countDocuments({
             shift: "day"
         }, function(err, count) {
             let day = count;
-            console.log("Day Shift: " + day);
-
             User.countDocuments({
                 shift: "night"
             }, function(err, count) {
                 let night = count;
-                console.log("Night Shift: " + night);
                 Log.countDocuments({
                     status: "active"
                 }, function(err, count) {
                     let active = count;
-                    console.log("Status active: " + active);
                     Log.countDocuments({
                         status: "logged out"
                     }, function(err, count) {
                         let done = count;
-                        console.log("Status done: " + done);
                         res.render("admin", {employees,day,night,active,done})
                     });
                 });
@@ -69,7 +62,6 @@ module.exports.logs_get = (req, res) => {
 
 module.exports.userrecords_get = async (req, res) => {
     const id = req.params.id;
-    console.log(id);
 
     const findUserRecord = await Record.find({user_id: id})
     const findUserName = await Record.findOne({user_id: id})
@@ -77,8 +69,8 @@ module.exports.userrecords_get = async (req, res) => {
     if(findUserRecord && findUserName){
         console.log(findUserRecord);
             res.render("user_records", {user_records: findUserRecord, name: findUserName.name})
-    } else{
-        consolge.log(err);
+    } else {
+        console.log("Error");
     }
 
 }
@@ -116,7 +108,7 @@ module.exports.users_get = (req, res) => {
 
 module.exports.userview_get = (req, res) => {
     const id = req.params.id;
-    console.log(id);
+    ;
 
     User.findById(id)
         .then(result => {
@@ -129,10 +121,10 @@ module.exports.userview_get = (req, res) => {
 
 module.exports.update_post = (req, res) => {
     const id = req.params.id;
-    console.log(id);
+    
 
     const user = new User(req.body);
-    console.log(user);
+    
 
     User.findByIdAndUpdate(id, {
             firstname: user.firstname,
@@ -155,19 +147,37 @@ module.exports.update_post = (req, res) => {
         })
 }
 
-module.exports.users_delete = (req, res) => {
+module.exports.users_delete = async (req, res) => {
     const id = req.params.id;
     console.log(id);
-    User.findByIdAndDelete(id)
-        .then(result => {
-            res.json({
-                redirect: "/users",
-                message: "User has been deleted successfully"
-            });
-        })
-        .catch(err => {
-            console.log(err);
-        })
+
+    const userCheck = await User.findOne({_id: id})
+    const findUserLog = await Log.findOne({user_id: id})
+    
+    const userIdLog = findUserLog.user_id;
+
+    if (userCheck && findUserLog == "" && findUserRecord == "") {
+        // delete User
+        await User.findByIdAndDelete(id)
+
+        return res.status(200).json({success: "User has been deleted successfully"})
+    }
+
+    if (userCheck && findUserLog) {
+        // delete User
+        await User.findByIdAndDelete(id)
+        
+
+        // delete User Logs
+        await Log.deleteOne({user_id: userIdLog})
+        
+
+        // delete User Records
+        await Record.deleteMany({user_id: userIdLog})
+        
+
+        return res.status(200).json({success: "User has been deleted successfully"})
+    }
 }
 
 module.exports.qrcodetester_get = (req, res) => {
